feat: Add QualifiedPreview component
